replace root jfieldvar by vars/jfieldvar
