+++ jcodemodel/src/main/java/com/helger/jcodemodel/vars/JFieldVar.java
@@ -1,20 +1,197 @@
+/*
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+ *
+ * Copyright (c) 1997-2010 Oracle and/or its affiliates. All rights reserved.
+ * Portions Copyright 2013-2026 Philip Helger + contributors
+ *
+ * The contents of this file are subject to the terms of either the GNU
+ * General Public License Version 2 only ("GPL") or the Common Development
+ * and Distribution License("CDDL") (collectively, the "License").  You
+ * may not use this file except in compliance with the License.  You can
+ * obtain a copy of the License at
+ * https://glassfish.dev.java.net/public/CDDL+GPL_1_1.html
+ * or packager/legal/LICENSE.txt.  See the License for the specific
+ * language governing permissions and limitations under the License.
+ *
+ * When distributing the software, include this License Header Notice in each
+ * file and include the License file at packager/legal/LICENSE.txt.
+ *
+ * GPL Classpath Exception:
+ * Oracle designates this particular file as subject to the "Classpath"
+ * exception as provided by Oracle in the GPL Version 2 section of the License
+ * file that accompanied this code.
+ *
+ * Modifications:
+ * If applicable, add the following below the License Header, with the fields
+ * enclosed by brackets [] replaced by your own identifying information:
+ * "Portions Copyright [year] [name of copyright owner]"
+ *
+ * Contributor(s):
+ * If you wish your version of this file to be governed by only the CDDL or
+ * only the GPL Version 2, indicate your decision by adding "[Contributor]
+ * elects to include this software in this distribution under the [CDDL or GPL
+ * Version 2] license."  If you don't indicate a single choice of license, a
+ * recipient has the option to distribute your version of this file under
+ * either the CDDL, the GPL Version 2 or to extend the choice of license to
+ * its licensees as provided above.  However, if you add GPL Version 2 code
+ * and therefore, elected the GPL Version 2 license, then the option applies
+ * only if the new code is made subject to such option by the copyright
+ * holder.
+ */
 package com.helger.jcodemodel.vars;
+
+import static com.helger.jcodemodel.util.JCHashCodeGenerator.getHashCode;
 
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
 
+import com.helger.base.enforce.ValueEnforcer;
+import com.helger.base.equals.EqualsHelper;
 import com.helger.jcodemodel.AbstractJType;
+import com.helger.jcodemodel.IJDocCommentable;
+import com.helger.jcodemodel.IJFormatter;
 import com.helger.jcodemodel.IVariableInitializer;
 import com.helger.jcodemodel.JDefinedClass;
+import com.helger.jcodemodel.JDocComment;
+import com.helger.jcodemodel.JFieldRef;
 import com.helger.jcodemodel.JMods;
+import com.helger.jcodemodel.modifiers.EMod;
+import com.helger.jcodemodel.writer.settings.Wrap.ListWrapping;
 
-public class JFieldVar extends com.helger.jcodemodel.JFieldVar
+/**
+ * A field that can have a {@link JDocComment} associated with it
+ */
+public class JFieldVar extends JBlockVar implements IJDocCommentable
 {
+  private final JDefinedClass m_aOwnerClass;
 
-  public JFieldVar (@NonNull JDefinedClass aOwnerClass, @NonNull JMods aMods, @NonNull AbstractJType aType, @NonNull String sName,
-      @Nullable IVariableInitializer aInit)
+  /**
+   * javadoc comments for this JFieldVar
+   */
+  private JDocComment m_aJavaDoc;
+
+  /**
+   * JFieldVar constructor
+   *
+   * @param aOwnerClass
+   *        The owning class.
+   * @param aMods
+   *        modifiers to use
+   * @param aType
+   *        Data type of this variable
+   * @param sName
+   *        Name of this variable
+   * @param aInit
+   *        Value to initialize this variable to
+   */
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  public JFieldVar (@NonNull final JDefinedClass aOwnerClass,
+                    @NonNull final JMods aMods,
+                    @NonNull final AbstractJType aType,
+                    @NonNull final String sName,
+                    @Nullable final IVariableInitializer aInit)
   {
-    super (aOwnerClass, aMods, aType, sName, aInit);
+    super (aMods, ValueEnforcer.notNull (aType, "type"), sName, aInit);
+    m_aOwnerClass = ValueEnforcer.notNull (aOwnerClass, "OwnerClass");
   }
 
+  /**
+   * @return The owning class. Never <code>null</code>.
+   */
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @NonNull
+  public JDefinedClass owner ()
+  {
+    return m_aOwnerClass;
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  public void name (@NonNull final String sNewName)
+  {
+    // make sure that the new name is available
+    ValueEnforcer.isFalse (m_aOwnerClass.containsField (sNewName),
+                           () -> "Field name '" + sNewName + "' is already in use");
+
+    final String sOldName = name ();
+    super.name (sNewName);
+    m_aOwnerClass.internalRenameField (sOldName, sNewName, this);
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  @NonNull
+  public JDocComment javadoc ()
+  {
+    if (m_aJavaDoc == null)
+    {
+      m_aJavaDoc = new JDocComment (m_aOwnerClass.owner ());
+    }
+    return m_aJavaDoc;
+  }
+
+  /**
+   * @return A field reference to this field variable. May be used for public static final
+   *         constants.
+   */
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @NonNull
+  public JFieldRef fieldRef ()
+  {
+    return new JFieldRef (m_aOwnerClass, this);
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  public void declare (@NonNull final IJFormatter f)
+  {
+    // Declaration
+    if (m_aJavaDoc != null)
+    {
+      f.generable (m_aJavaDoc);
+    }
+    super.declare (f);
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  public boolean equals (final Object o)
+  {
+    if (o == this)
+      return true;
+    if (!super.equals (o))
+      return false;
+    final JFieldVar rhs = (JFieldVar) o;
+    return EqualsHelper.equals (m_aOwnerClass, rhs.m_aOwnerClass);
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  public int hashCode ()
+  {
+    return getHashCode (super.hashCode (), new Object [] { m_aOwnerClass });
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  public JFieldVar emod (EMod emod, EMod... emods)
+  {
+    mods ().emod (EMod.ALLOWED_FIELD, emod, emods);
+    return this;
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  public JFieldVar removeEMod (EMod... emods)
+  {
+    mods ().removeEMod (emods);
+    return this;
+  }
+
+  @Deprecated(forRemoval = true, since = "4.3.0")
+  @Override
+  protected ListWrapping extractWrappingOptions (@NonNull IJFormatter f)
+  {
+    return f.settings ().wrap.variables.field;
+  }
 }
