+++ jcodemodel/src/test/java/com/helger/jcodemodel/AnnotationUseFuncTest.java
@@ -50,6 +50,7 @@ import org.junit.Test;
 
 import com.helger.jcodemodel.exceptions.JCodeModelException;
 import com.helger.jcodemodel.util.CodeModelTestsHelper;
+import com.helger.jcodemodel.vars.JFieldVar;
 
 /**
  * A test program for the annotation use features Note: Not all the generated code would make sense
@@ -169,6 +170,8 @@ public final class AnnotationUseFuncTest
 
     CodeModelTestsHelper.parseCodeModel (cm);
     if (false)
+    {
       CodeModelTestsHelper.compileCodeModel (cm);
+    }
   }
 }
