+++ jcodemodel/src/test/java/com/helger/jcodemodel/JFieldRefTest.java
@@ -43,6 +43,7 @@ package com.helger.jcodemodel;
 import org.junit.Test;
 
 import com.helger.jcodemodel.util.CodeModelTestsHelper;
+import com.helger.jcodemodel.vars.JFieldVar;
 
 /**
  * Test class for class {@link JFieldRef}
