+++ jcodemodel/src/test/java/com/helger/jcodemodel/supplementary/issues/Issue30FuncTest.java
@@ -47,11 +47,11 @@ import com.helger.jcodemodel.JCodeModel;
 import com.helger.jcodemodel.JDefinedClass;
 import com.helger.jcodemodel.JDirectClass;
 import com.helger.jcodemodel.JExpr;
-import com.helger.jcodemodel.JFieldVar;
 import com.helger.jcodemodel.JMethod;
 import com.helger.jcodemodel.JMod;
 import com.helger.jcodemodel.JPackage;
 import com.helger.jcodemodel.util.CodeModelTestsHelper;
+import com.helger.jcodemodel.vars.JFieldVar;
 
 /**
  * Test for https://github.com/phax/jcodemodel/issues/30
@@ -88,7 +88,9 @@ public final class Issue30FuncTest
     CodeModelTestsHelper.parseCodeModel (cm);
     // Won't work
     if (false)
+    {
       CodeModelTestsHelper.compileCodeModel (cm);
+    }
   }
 
   @Test
@@ -112,7 +114,9 @@ public final class Issue30FuncTest
     CodeModelTestsHelper.parseCodeModel (cm);
     // Won't work
     if (false)
+    {
       CodeModelTestsHelper.compileCodeModel (cm);
+    }
   }
 
   @Test
@@ -137,6 +141,8 @@ public final class Issue30FuncTest
     CodeModelTestsHelper.parseCodeModel (cm);
     // Won't work
     if (false)
+    {
       CodeModelTestsHelper.compileCodeModel (cm);
+    }
   }
 }
