+++ jcodemodeltests/src/main/java/com/helger/jcodemodel/tests/arrayinit/ArrayInitTestGen.java
@@ -16,11 +16,11 @@ package com.helger.jcodemodel.tests.arrayinit;
 
 import com.helger.jcodemodel.JDefinedClass;
 import com.helger.jcodemodel.JExpr;
-import com.helger.jcodemodel.JFieldVar;
 import com.helger.jcodemodel.JMod;
 import com.helger.jcodemodel.JPackage;
 import com.helger.jcodemodel.compile.annotation.TestJCM;
 import com.helger.jcodemodel.exceptions.JCodeModelException;
+import com.helger.jcodemodel.vars.JFieldVar;
 
 @TestJCM
